+++ gtk/gtkcolorutils.h
@@ -40,11 +40,11 @@
 G_BEGIN_DECLS
 
 GDK_AVAILABLE_IN_ALL
-void gtk_hsv_to_rgb (gdouble  h, gdouble  s, gdouble  v,
-                     gdouble *r, gdouble *g, gdouble *b);
+void gtk_hsv_to_rgb (float  h, float  s, float  v,
+                     float *r, float *g, float *b);
 GDK_AVAILABLE_IN_ALL
-void gtk_rgb_to_hsv (gdouble  r, gdouble  g, gdouble  b,
-                     gdouble *h, gdouble *s, gdouble *v);
+void gtk_rgb_to_hsv (float  r, float  g, float  b,
+                     float *h, float *s, float *v);
 
 G_END_DECLS
